add improvements, add tests
'formatter' is a QueryParser attribute and it's a string now.
There's no need to pass it to get_data function as a separate argument.
Changed the return type (inside the docstring in format_structure function)
from 'str' to Union[Dict[str, Any], List[str]].
Added a few more digits to re pattern to allow to find the new metrics.
Added simple unit tests.